+++ src/main/java/fi/dy/masa/malilib/network/IPluginClientPlayHandler.java
@@ -3,8 +3,8 @@ package fi.dy.masa.malilib.network;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.Objects;
-import net.neoforged.neoforge.network.handling.IPayloadContext;
-import net.neoforged.neoforge.network.handling.IPayloadHandler;
+import org.thinkingstudio.fabric.api.client.networking.v1.ClientPlayNetworking;
+import org.thinkingstudio.fabric.api.networking.v1.PayloadTypeRegistry;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import net.minecraft.client.network.ClientPlayNetworkHandler;
 import net.minecraft.nbt.NbtCompound;
@@ -16,13 +16,12 @@ import net.minecraft.network.packet.Packet;
 import net.minecraft.network.packet.c2s.common.CustomPayloadC2SPacket;
 import net.minecraft.util.Identifier;
 import fi.dy.masa.malilib.MaLiLib;
-import org.thinkingstudio.mafglib.util.NeoNetwork;
 
 /**
  * Interface for ClientPlayHandler, for downstream mods.
  * @param <T> (Payload)
  */
-public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayloadHandler<T>
+public interface IPluginClientPlayHandler<T extends CustomPayload> extends ClientPlayNetworking.PlayPayloadHandler<T>
 {
     int FROM_SERVER = 1;
     int TO_SERVER = 2;
@@ -57,8 +56,8 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
     void reset(Identifier channel);
 
     /**
-     * Register your Payload with NeoForge network.
-     * See the NeoForge's network Java Docs under PayloadTypeRegistry -> register()
+     * Register your Payload with Fabric API.
+     * See the fabric-networking-api-v1 Java Docs under PayloadTypeRegistry -> register()
      * for more information on how to do this.
      * -
      * @param id (Your Payload Id<T>)
@@ -73,12 +72,12 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
             {
                 switch (direction)
                 {
-                    case TO_SERVER, FROM_CLIENT -> NeoNetwork.getPayloadRegistrar().playToServer(id, codec, this);
-                    case FROM_SERVER, TO_CLIENT -> NeoNetwork.getPayloadRegistrar().playToClient(id, codec, this);
+                    case TO_SERVER, FROM_CLIENT -> PayloadTypeRegistry.playC2S().register(id, codec);
+                    case FROM_SERVER, TO_CLIENT -> PayloadTypeRegistry.playS2C().register(id, codec);
                     default ->
                     {
-                        NeoNetwork.getPayloadRegistrar().playToServer(id, codec, this);
-                        NeoNetwork.getPayloadRegistrar().playToClient(id, codec, this);
+                        PayloadTypeRegistry.playC2S().register(id, codec);
+                        PayloadTypeRegistry.playS2C().register(id, codec);
                     }
                 }
             }
@@ -97,22 +96,20 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
     /**
      * Register your Packet Receiver function.
      * You can use the HANDLER itself (Singleton method), or any other class that you choose.
-     * See the NeoForge's network Java Docs under PayloadRegistrar.playToClient()
+     * See the fabric-network-api-v1 Java Docs under ClientPlayNetworking.registerGlobalReceiver()
      * for more information on how to do this.
      * -
      * @param id (Your Payload Id<T>)
-     * @param codec (Your Payload's CODEC)
      * @param receiver (Your Packet Receiver // if null, uses this::receivePlayPayload)
      * @return (True / False)
      */
-    default boolean registerPlayReceiver(@Nonnull CustomPayload.Id<T> id, @Nonnull PacketCodec<? super RegistryByteBuf,T> codec, @Nullable IPayloadHandler<T> receiver)
+    default boolean registerPlayReceiver(@Nonnull CustomPayload.Id<T> id, @Nullable ClientPlayNetworking.PlayPayloadHandler<T> receiver)
     {
         if (this.isPlayRegistered(this.getPayloadChannel()))
         {
             try
             {
-                NeoNetwork.getPayloadRegistrar().playToClient(id, codec, Objects.requireNonNullElse(receiver, this::receivePlayPayload));
-                return true;
+                return ClientPlayNetworking.registerGlobalReceiver(id, Objects.requireNonNullElse(receiver, this::receivePlayPayload));
             }
             catch (IllegalArgumentException e)
             {
@@ -125,16 +122,16 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
         return false;
     }
 
-//    /**
-//     * Unregisters your Packet Receiver function.
-//     * You can use the HANDLER itself (Singleton method), or any other class that you choose.
-//     * See the fabric-network-api-v1 Java Docs under ClientPlayNetworking.unregisterGlobalReceiver()
-//     * for more information on how to do this.
-//     */
-//    default void unregisterPlayReceiver()
-//    {
-//        ClientPlayNetworking.unregisterGlobalReceiver(this.getPayloadChannel());
-//    }
+    /**
+     * Unregisters your Packet Receiver function.
+     * You can use the HANDLER itself (Singleton method), or any other class that you choose.
+     * See the fabric-network-api-v1 Java Docs under ClientPlayNetworking.unregisterGlobalReceiver()
+     * for more information on how to do this.
+     */
+    default void unregisterPlayReceiver()
+    {
+        ClientPlayNetworking.unregisterGlobalReceiver(this.getPayloadChannel());
+    }
 
     /**
      * Receive Payload by pointing static receive() method to this to convert Payload to its data decode() function.
@@ -142,7 +139,7 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
      * @param payload (Payload to decode)
      * @param ctx (Fabric Context)
      */
-    void receivePlayPayload(T payload, IPayloadContext ctx);
+    void receivePlayPayload(T payload, ClientPlayNetworking.Context ctx);
 
     /**
      * Receive Payload via the legacy "onCustomPayload" from a Network Handler Mixin interface.
@@ -165,6 +162,7 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
     default void decodeNbtCompound(Identifier channel, NbtCompound data) {}
     default void decodeByteBuf(Identifier channel, MaLiLibBuf data) {}
     default <D> void decodeObject(Identifier channel, D data1) {}
+    default <P extends IClientPayloadData> void decodeClientData(Identifier channel, P data) {}
 
     /**
      * Payload Encoder wrapper function.
@@ -175,6 +173,7 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
     default void encodeNbtCompound(NbtCompound data) {}
     default void encodeByteBuf(MaLiLibBuf data) {}
     default <D> void encodeObject(D data1) {}
+    default <P extends IClientPayloadData> void encodeClientData(P data) {}
 
     /**
      * Used as an iterative "wrapper" for Payload Splitter to send individual Packets
@@ -184,7 +183,7 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
     void encodeWithSplitter(PacketByteBuf buf, ClientPlayNetworkHandler handler);
 
     /**
-     * Sends the Payload to the server using the NeoForge's network interface.
+     * Sends the Payload to the server using the Fabric-API interface.
      * -
      * @param payload (The Payload to send)
      * @return (true/false --> for error control)
@@ -193,9 +192,9 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
     {
         if (payload.getId().id().equals(this.getPayloadChannel()) && this.isPlayRegistered(this.getPayloadChannel()))
         {
-            if (NeoNetwork.canSend(payload.getId()))
+            if (ClientPlayNetworking.canSend(payload.getId()))
             {
-                NeoNetwork.sendToServer(payload);
+                ClientPlayNetworking.send(payload);
                 return true;
             }
         }