+++ src/main/java/org/thinkingstudio/mafglib/MaFgLib.java
@@ -4,18 +4,15 @@ import fi.dy.masa.malilib.MaLiLib;
 import fi.dy.masa.malilib.MaLiLibConfigGui;
 import fi.dy.masa.malilib.MaLiLibReference;
 import net.neoforged.api.distmarker.Dist;
-import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.loading.FMLLoader;
-import org.thinkingstudio.mafglib.util.NeoNetwork;
 import org.thinkingstudio.mafglib.util.NeoUtils;
 
-@Mod(MaLiLibReference.MOD_ID)
+@Mod(value = MaLiLibReference.MOD_ID, dist = Dist.CLIENT)
 public class MaFgLib {
-    public MaFgLib(IEventBus modEventBus) {
+    public MaFgLib() {
         if (FMLLoader.getDist() == Dist.CLIENT) {
             MaLiLib.onInitialize();
-            NeoNetwork.init(modEventBus);
 
             // Config Screen
             NeoUtils.getInstance().registerModConfigScreen(MaLiLibReference.MOD_ID, (screen) -> {