+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ## Change
-- fix network issue (#48, #45, #47)
+- use NeoNetwork replace ForgedNetworkingAPI and NeoForge native network api